sunxi: add initial Allwinner H6 support
Allwinner H6 is a new SoC in the sun50i series, with Cortex-A53 cores
and refactored memory map/CCU.

Add initial support for it, including initial MMC support.

The offset of the environment is moved to 0xe0000, which used to be the
space reserved for Falcon mode in legacy u-boot-sunxi. This change will
prevent the U-Boot space issue which happened on several sunxi boards.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>